+++ term/src/terminal.rs
@@ -91,7 +91,7 @@ impl DerefMut for Terminal {
     }
 }
 
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, FromDynamic, ToDynamic)]
 #[cfg_attr(feature = "use_serde", derive(Serialize, Deserialize))]
 pub struct TerminalSize {
     pub rows: usize,